Changed long to ptrdiff_t fix 32bit build

+++ src/dlangui/platforms/sdl/sdlapp.d
@@ -103,7 +103,7 @@ class SDLWindow : Window {
         debug Log.d("Destroying SDL window");
         
         if (_parent) {
-            long index = countUntil(_parent._children,this);
+            ptrdiff_t index = countUntil(_parent._children,this);
             if (index > -1 ) {
                 _parent._children=_parent._children.remove(index);
             }
